+++ lib/__init__.py
@@ -22,24 +22,3 @@ class Keys:
 
 
 
-class Context:
-    def __init__(self, *args):
-        if len(args) == 1:
-            arg = args[0]
-            if type(arg) == str:
-                self.GALAXY_SERVER, self.API_KEY, self.KUBECONFIG = parse_profile(arg)
-            elif type(arg) == dict:
-                self.GALAXY_SERVER = arg['GALAXY_SERVER']
-                self.API_KEY = arg['API_KEY']
-                self.KUBECONFIG = arg['KUBECONFIG']
-            else:
-                raise Exception(f'Invalid arg for Context: {type(arg)}')
-        elif len(args) == 3:
-            self.GALAXY_SERVER = args[0]
-            self.API_KEY = args[1]
-            self.KUBECONFIG = args[2]
-        else:
-            raise Exception(f'Invalid args for Context. Expected one or three, found {len(args)}')
-
-
-
