+++ lib/job.py
@@ -1,6 +1,5 @@
 import json
-from lib import Context
-from .common import connect
+from .common import connect, Context
 from pprint import pprint
 import logging
